pic24f: Generate scripts appropriate for use with bootloaders
This shell script will generate linker scripts which can be used with
bootloaders.  The design is that the same linker script can be used for
both the bootloader and for the application to be loaded by the bootloader
and uses macros to make this process more robust and less error-prone.
Preprocessor definitions are used to specify which mode (bootloader
or application) is being used for a particular build.

This implementation includes re-mapping the interrupt vector table (IVT).
The idea is that the entries in the IVT point to a map which contains a
GOTO instruction for each interrupt which jumps to the actual interrupt
handler.  The IVT is filled in by the bootloader's linker script, and the
GOTO table is filled in by the application's linker script.  The tricky
part is that the bootloader has to know where each entry is in the GOTO
map even though the application actually fills this table.  This is done
by a macro which creates symbols ([INTERRUPT_NAME]_ADDR) and increments
the address pointer (the '.') for each interrupt (look for X##_ADDR in the
output).

The basic idea is shown pictorally below:

                 |  bootloader |        application           |
interrupt event -> IVT entry  -> goto table -> applicaiton ISR